requirements.txt
flask == 2.3.3
flask-mysqldb == 1.0.1
PyJWT == 2.8.0
flask-cors